Adding variable bin histograms and fixing compiler warnings
Adding two options for variable bin histograms.  One uses the
traditional root style of an array of pointers; the other takes a vector
and decides on the number of bins based on the size of the vector.  Also
fixing some compiler warnings; it's not beautiful, but it does make the
compilation nicer.